+++ components/plurals/src/provider.rs
@@ -15,7 +15,7 @@
 //!
 //! Read more about data providers: [`icu_provider`]
 
-use crate::rules::runtime::ast::Rule;
+use crate::provider::rules::runtime::ast::Rule;
 use crate::{PluralCategory, PluralElements, PluralElementsInner, PluralOperands, PluralRules};
 use alloc::borrow::{Cow, ToOwned};
 use alloc::boxed::Box;
