+++ components/plurals/src/provider/rules/reference/ast.rs
@@ -130,7 +130,7 @@ pub struct Condition(pub Vec<AndCondition>);
 /// # Examples
 ///
 /// All AST nodes can be built explicitly, as seen in the example. However, due to its complexity, it is preferred to build the
-/// AST using the [`parse()`](crate::rules::reference::parser::parse()) function.
+/// AST using the [`parse()`](crate::provider::rules::reference::parser::parse()) function.
 ///
 /// ```text
 /// "i = 3 and v = 0"
@@ -171,7 +171,7 @@ pub struct AndCondition(pub Vec<Relation>);
 /// # Examples
 ///
 /// All AST nodes can be built explicitly, as seen in the example. However, due to its complexity, it is preferred to build the
-/// AST using the [`parse()`](crate::rules::reference::parser::parse()) function.
+/// AST using the [`parse()`](crate::provider::rules::reference::parser::parse()) function.
 ///
 /// ```text
 /// "i = 3"
@@ -229,7 +229,7 @@ pub enum Operator {
 /// # Examples
 ///
 /// All AST nodes can be built explicitly, as seen in the example. However, due to its complexity, it is preferred to build the
-/// AST using the [`parse()`](crate::rules::reference::parser::parse()) function.
+/// AST using the [`parse()`](crate::provider::rules::reference::parser::parse()) function.
 ///
 /// ```text
 /// "i % 100"
@@ -260,7 +260,7 @@ pub struct Expression {
 /// # Examples
 ///
 /// All AST nodes can be built explicitly, as seen in the example. However, due to its complexity, it is preferred to build the
-/// AST using the [`parse()`](crate::rules::reference::parser::parse()) function.
+/// AST using the [`parse()`](crate::provider::rules::reference::parser::parse()) function.
 ///
 /// ```text
 /// "i"
@@ -300,7 +300,7 @@ pub enum Operand {
 /// # Examples
 ///
 /// All AST nodes can be built explicitly, as seen in the example. However, due to its complexity, it is preferred to build the
-/// AST using the [`parse()`](crate::rules::reference::parser::parse()) function.
+/// AST using the [`parse()`](crate::provider::rules::reference::parser::parse()) function.
 ///
 /// ```text
 /// "5, 7, 9"
@@ -357,7 +357,7 @@ pub enum RangeListItem {
 /// # Examples
 ///
 /// All AST nodes can be built explicitly, as seen in the example. However, due to its complexity, it is preferred to build the
-/// AST using the [`parse()`](crate::rules::reference::parser::parse()) function.
+/// AST using the [`parse()`](crate::provider::rules::reference::parser::parse()) function.
 ///
 /// ```text
 /// "99"
